requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-filter-stubs==0.1.2
django-stubs==1.13.0
django-stubs-ext==0.7.0
django-tables2==2.4.1
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
dnspython==2.2.1
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.2
flake8==5.0.4
freezegun==1.2.2
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-dj-database-url==1.0.0
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.12
validators==0.20.0
whitenoise==6.2.0
